lib: refacto sclangexpbindings to a folder + move Env inside + comment Some SynthDefs code
