+++ renardo_lib/renardo_lib/ServerManager/__init__.py
@@ -529,12 +529,12 @@ class ServerManager:
 
         return
 
-    def loadCompiled(self, fn):
-        """ Sends a message to SuperCollider to load a compiled SynthDef file """
-        msg = OSCMessage()
-        msg.setAddress('/d_load')
-        msg.append(fn)
-        self.client.send(msg)
+    #def loadCompiled(self, fn):
+    #    """ Sends a message to SuperCollider to load a compiled SynthDef file """
+    #    msg = OSCMessage()
+    #    msg.setAddress('/d_load')
+    #    msg.append(fn)
+    #    self.client.send(msg)
 
     def dumpOSC(self, value=1):
         """ Debug - Dumps OSC messages SCLang side """
